\name{fish.data
}
\alias{fish.data
}
\title{fish.data
}
\description{fish.data is an example data set to fit models for total abundance (N) and Species Richness (S)
}
\details{A much reduced data set from CSIRO voyages in the 1970s
}
\section{Authors}{Piers Dunstan and Scott Foster
}
\keyword{misc}
