% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggplot_AnomalyDetection.rpca}
\alias{ggplot_AnomalyDetection.rpca}
\title{ggplot for AnomalyDetection}
\usage{
ggplot_AnomalyDetection.rpca(anomalyDetection)
}
\arguments{
\item{anomalyDetection}{output from AnomalyDetection.rpca}
}
\description{
ggplot function which shows the low rank signal in blue, the random noise in green,
and any outliers in red. If a transformation was applied, these signals will be plotted
in the transformed space, along with the original time series
}
\examples{
frequency = 7
numPeriods = 10
ts.sinusoidal = sin((2 * pi / frequency ) * 1:(numPeriods * frequency))
ts = ts.sinusoidal
ts = sin((2 * pi / frequency ) * 1:(numPeriods * frequency))
ts[58:60] = 100
ggplot_AnomalyDetection.rpca(AnomalyDetection.rpca(ts)) + ggplot2::theme_grey(base_size = 25)
}

