\name{RAD}
\alias{RAD}
\title{RAD}
\description{
The package RAD is used to fit models to Rank Abundance Distributions, and predict those distributions using new covariates. Details are provided in Foster and Dunstan (2009) and Dunstan and Foster (some day).

negBinMod fits models for N, truncMod fits models for S and MDMnMod fits models for nij
}
\references{
Foster, S.D. and Dunstan, P.K. (2009) The Analysis of Biodiversity Using Rank Abundance Distributions. Biometrics DOI: 10.1111/j.1541-0420.2009.01263.x

Dunstan, P.K. and Foster, S.D. (some day) RAD Biodiversity: Prediction of Rank Abundance Distributions from Deep Water Benthic Assemblages.
}
\keyword{misc}
