% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RcppRPCA}
\alias{RcppRPCA}
\title{Robust Principal Component Pursuit}
\usage{
RcppRPCA(X, Lpenalty = -1, Spenalty = -1, verbose = FALSE)
}
\arguments{
\item{X}{numeric matrix}

\item{Lpenalty}{scalar to penalize singular values.
A default of -1 is used as a sentinel value. Under this sentinel value
a smart thresholding algorithm sets the value of Lpenalty, so the user does not need to.}

\item{Spenalty}{scalar to penalize remainder matrix to find anomalous values.
A default of -1 is used as a sentinel value. Under this sentinel value
a smart thresholding algorithm setes the value of Spenalty, so the user does not need to.}
}
\value{
a list with 4 matrices.  X is decomposed into L + S + E where L is low rank,
S is sparse and E is the remainder matrix of noise
}
\description{
Robust Principal Component Pursuit
}

