% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RcppSVT}
\alias{RcppSVT}
\title{Singular Value Thresholding on a numeric matrix}
\usage{
RcppSVT(X, penalty)
}
\arguments{
\item{X}{numeric matrix}

\item{penalty}{scalar to penalize singular values}
}
\value{
a list with 2 components, the singular value thresholded matrix and its thresholded singular values
}
\description{
Singular Value Thresholding on a numeric matrix
}

